pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-tagger"
version = "0.1.0"
description = "Cross-platform GUI app and library that tags and captions images using multiple AI models."
readme = "README.md"
authors = [
  { name = "Your Name", email = "you@example.com" }
]
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
  "pillow>=10.2",
  "piexif>=1.1.3",
  "PyYAML>=6.0",
  "pydantic>=2.6",
  "pyside6>=6.6",
  "tqdm>=4.66",
  "numpy>=1.24",
  "typing_extensions>=4.8"
]

[project.optional-dependencies]
clip = ["open-clip-torch>=2.24", "torch"]
transformers = ["transformers", "accelerate", "sentencepiece"]
onnx = ["onnxruntime", "onnx"]

[project.urls]
Homepage = "https://example.com/image-tagger"
Repository = "https://example.com/image-tagger.git"

[project.scripts]
image-tagger = "image_tagger.__main__:main"

[tool.coverage.run]
source = ["image_tagger"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]

